Fix error on column names with underscores.
The allowed fields are formatted and collected in
`formatted_allowed_fields` array, which is used during the comparison.
During comparison though, keys from the `params[:data]` hash are not
formatted before verifying whether they are allowed. Thus keys with
underscores will always fail. This fix makes sure the comparison is done
using the formatted version of the keys.