Save after 2nd callback in Resource#change
Instead of saving directly after the yield in the second run_callbacks
call, save after the second run_callbacks is finished.

This allows for using callbacks like after_replace_fields to work with
the model with the newly set attributes before it gets saved. This is
useful when doing authorization before creation, as well as using
non-datastore persisted attributes on the model to set specific data
on the model before it gets persisted.