Only loop through filters if #each is implemented
In Request#parse_filters, only loop through the filters if the passed in
value implements the #each method. Without this check,

    NoMethodError: undefined method `each' for "all":String

gets thrown when request filters are not in a not valid format.

Adds the INVALID_FILTERS_SYNTAX error to let the requester know that the
filters syntax was invalid.