Provide filter context to resource records query
Sometimes when overriding the records method on a resource it is useful
to have access to the current filters. For instance if the underlying 
base query for the resource changes based on a filter param. This can 
not be done in the apply_filter method because compound filters would 
produce inconsistent results depending on the order in which they are 
run.