Add polymorphic support to has_many relationships.
This adds support for the serialization of `has_many` relationships. I'm not sure if it handles every situation that people are imagining but it addresses the primary issue that I saw which was the need to return the correct type for the individual resources. As with the solution for polymorphic `has_one` relationships, you need to add a `Resource` and `Controller` class for each potential type. I didn't address anything new related to creating or updating `has_many` relationships in this commit.