Add some additional info about raising errors
This addresses the subject of raising errors with all the relevant
info that a user would need to know how to handle their own application
level errors.

I placed it right after the first mention of raising custom errors, but
I'm happy to move it elsewhere if that would be better.