Handle namespaced models
Fix for #523 (possible fix for #513)

Resource lookup was failing unless the model namespace was the same as
the resource namespace, excluding the module path. So while it worked
for the resource, `/api/v2/posts` having a model `Post`. It didn't
reliably work for a model `Legacy::Post`.

Overriding `resource_for_type` did solve part of the problem, but when
the mismatched resource was a relationship, `module_path` was wrong
and the resource could not be found.

I tried creating `/api/v2/legacy/posts` as a copy of `/api/v2/posts` and
this got me further, but still there were some cases were the resource
could not be found. (So far I am unable to generalize these cases, as
I could see no differences between these and the ones where this
technique worked.)

I had to resort to storing the `model` and `path` in the
`@@resource_types` class variable. And use that information in the few
`resource_for` calls that tried to lookup by model.