fix bug with server_error_callbacks being called from subclasses
Wow, this one took me a while to figure out.

There were a couple of changes made to the ActsAsResourceController.
One of the changes switched away from the ActiveSupport::Concern class_methods
approach, and one of them refactored the way that the server_error_callbacks
were stored. The problem was that the class instance variable wasn't visible
to subclasses and the test suite missed it because it tested controller classes
that had the on_server_error block set directly on them. I fixed that problem
and completed the switch back to the non-ActiveSupport::Concern approach.