Make fetchable_fields more like other field overrides
Adds a class-level `self.fetchable_fields(context)` method to Resource
and changes the existing `Resource#fetchable_fields` to call the class-
level version with the current context. Updates docs to encourage 
overwriting `self.fetchable_fields` to be parallel with the workflows
for `updatable_fields`, `creatable_fields`, and `sortable_fields`.

This change should be backwards-compatible with existing instance-level
overrides of `fetchable_fields`.