error serializing relationships w same foreign key
when multiple relationships share the same foreign key and they are both
`included` the relationships are correctly loaded from the DB but
when serialized to json the records for the first defined relationship
are used to populate all the others.

for example:

given the following

class Thing < ActiveRecord::Base
  belongs_to :example
end

class Example < ActiveRecord::Base
  has_many :things

  has_many :cool_things, -> { where(cool: true) },
  foreign_key: :example_id, class_name: 'Thing'

  has_many :uncool_things, -> { where(cool: false) },
  foreign_key: :example_id, class_name: 'Thing'
end

class ExampleResource < JSONAPI::Resource
  has_many :cool_things, foreign_key: :example_id,
  relation_name: :cool_things, class_name: 'Thing

  has_many :uncool_things, foreign_key: :example_id,
  relation_name: :cool_things, class_name: 'Thing'
end

when a request is made for an example `include`ing its `cool_things` and
its `uncool_things` the `cool_things` records will be used to populate
the `uncool_things` as well.

this commit contains a test demonstrating this issue.