patch to serialize relationships w same foreign key
instead of fetching records using the foreign_key use the relationship
name instead. the value stored in the map under the foreign key contains
the records for the first relationship defined using that foreign
key. by contrast using the relationship name returns the records fetched
for the relationship defined with that name.