Make default sort criteria blank.
Enable `records(options)` to know if a sort criteria was passed
by the client.

Before this change, we didn't know if the sort criteria "id ASC" passed in
the `options` hash was provided by the client or was the default one.
This would prevent us from implementing a default sort order.