Get relationship id from the resource
We should avoid getting relationship id from the `model` directly to allow clients to customize how relationship ids are fetched. The resource exposes the relationship id through a method which delegates to the `model`. This way we allow clients to override the relationship id in the resource to fit their needs.