Rework model_name in derived resources
Fixes an issue where the recreated relationships might be using the wrong model_name if the model_name is changed in a resource that is derived from a non abstract resource, such as done in many of the tests.
Breaking change: Derived resources now use the model name of their base resource, if it is not abstract.
Cleans up the tests to not have the warnings about missing models.
Fixes issue in _model_class related to #952, and renames @model (class level) to @model_class to avoid confusion with instance level @model

Fixes #952, #654