Alters build_joins sql for has_one relationships
The existing tests are passing because they seem to all use a self-
referential `has_one` relationship. If the foreign_key is on a different
table, however, the queries will fail. So if it the relationship is a
has_one, we need to alter the query to look for the foreign-key on the
related table.