Use controller for route helpers in LinkBuilder
When `Rails.application.routes.url_helpers` is called a new module is
created. This is expensive both in cpu and memory resources.

In order to avoid storing the url_helpers module as a Singleton these
changes pass the controller, which has included a copy of this module,
through the ResourceSerializer and the LinkBuilder.

For unit testing we fall back to using the module from
`TestApp.routes.url_helpers` since a controller is not available.